requirements/requirements.txt
Flask==1.1.1
python-dotenv==0.11.0
redis==3.4.1
slackclient==1.3.2
slackeventsapi==2.1.0
oauth2client==4.1.3
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
